New Dockerfile which uses Alpine
The previous Dockerfile (based on CentOS) was not working seamlessly.
There were a number of steps you would need to take after cloning the
repo in order to get it working (which weren't documented).
The image size was also quite large, (close to 800MB on my system),
which seemed excessive for building/testing a MQTT interface.

This change addresses the above in the following ways:

- Dockerfile in root of repo, can be built using standard docker build
  commands, without any changes required.

- Base OS changed from CentOS to Alpine

- Restructured the Dockerfile somewhat to cut down on ADD/RUN commands
  (which contribute to the image bloat)

- Faster build time, as a result of the lighter-weight OS